Extract image composition backend module (#1272)
