+++ app/index.ts
@@ -28,6 +28,7 @@ import IGeesomeDriversModule from "./modules/drivers/interface.js";
 import IGeesomeContentModule from "./modules/content/interface.js";
 import IGeesomeInviteModule from "./modules/invite/interface.js";
 import IGeesomeGroupModule from "./modules/group/interface.js";
+import IGeesomeImageCompositionModule from "./modules/imageComposition/interface.js";
 import IGeesomeApiModule from "./modules/api/interface.js";
 import {IGeesomeApp, IUserInput} from "./interface.js";
 import {GeesomeEmitter} from "./events.js";
@@ -137,6 +138,7 @@ function getModule(config, appPass) {
       staticId: IGeesomeStaticIdModule,
       invite: IGeesomeInviteModule,
       group: IGeesomeGroupModule,
+      imageComposition: IGeesomeImageCompositionModule,
       accountStorage: IGeesomeAccountStorageModule,
       communicator: IGeesomeCommunicatorModule,
       storage: IGeesomeStorageModule,
