+++ app/modules/imageComposition/models/imageCompositionOperation.ts
