+++ app/modules/imageComposition/models/index.ts
@@ -1,0 +1,6 @@
+import type {Sequelize} from 'sequelize';
+
+export default async function (sequelize: Sequelize, appModels) {
+	appModels.ImageCompositionOperation = await (await import('./imageCompositionOperation.js')).default(sequelize, appModels);
+	return appModels;
+}
