+++ docs/database-scalability-inventory.md
@@ -52,7 +52,7 @@ This file is generated by `npm run database:scalability:update`. Do not hand-edi
 | Large fixture benchmark | database:scalability:fixture / explain / generated-output | seeds a 100k-post group fixture plus category, mixed image/text/json content metadata, attachment, quota, preview, and static-ID side rows, then emits current EXPLAIN ANALYZE probes and generated-output pressure signals against fixture/restored groups selected by name or id; optional category/content/static-ID probes are skipped when restored data lacks sample rows | future index/query changes and generated-output body/copy pressure can be measured against production-scale fixture or a specific restored group id; generated reports remain intentionally uncommitted | `check/databaseScalabilityFixture.ts` |
 | Generated output pressure | database:scalability:generated-output | reports selected-post attachment rows, unique text/json body read candidates, RSS selected text reads, non-text storage copy candidates, and group-manifest inline/index pressure from a fixture/restored group selected by name or id | the large-output follow-up can distinguish persisted-snippet pressure from cache tuning, storage-copy pressure, and large-manifest cutoff pressure before changing runtime defaults or adding schema | `check/databaseGeneratedOutputPressure.ts` |
 | Timeline API | getGroupPosts | keyset/cursor and legacy offset paths select page post IDs first, hydrate contents/reposts for the bounded page, and allow offset callers to skip totals | page selection avoids attachment joins; cursor pages and includeTotal=false skip counts, while default legacy offset callers can still request totals | `app/modules/group/index.ts` |
-| Image composition timeline | getImageCompositions | filters by the dedicated post type before page hydration and uses the group/type/status/publishedAt/id cursor index | composition pages remain bounded and avoid propertiesJson scans; legacy offset callers can still request totals | `app/modules/group/index.ts + app/modules/group/models/post.ts` |
+| Image composition timeline | getImageCompositions | filters by the dedicated post type before page hydration and uses the group/type/status/publishedAt/id cursor index | composition pages remain bounded and avoid propertiesJson scans; legacy offset callers can still request totals | `app/modules/imageComposition/index.ts + app/modules/group/models/post.ts` |
 | Global post listings | getAllPosts | selects cursor-aware page post refs first, then hydrates contents for the bounded page; exports can iterate lightweight refs in cursor batches | content joins no longer drive page selection; high-volume all-post scans can avoid offsets through cursor ref batches | `app/modules/group/index.ts` |
 | Unread counters | getGroupUnreadPostsData | count posts newer than the stored (readAt, readPostId) cursor when present, with readAt-only fallback | count is backed by the timeline index and can disambiguate same-timestamp posts; old readAt-only rows still use the legacy timestamp predicate | `app/modules/group/index.ts` |
 | Group manifest generation | generateGroupManifest | loads the previous posts trie, scans changed/deleted lightweight post refs in (updatedAt,id) cursor batches, then unsets deleted local IDs | content/repost hydration, timestamp-only watermarks, and large changed-ref windows are avoided; large groups can avoid the legacy inline trie when enabled, while default chunked-only rollout and page-level incremental updates still need follow-up | `app/modules/entityJsonManifest/index.ts` |
